+++ apps/buildog/src/app/(auth)/layout.tsx
@@ -1,13 +1,12 @@
-import { Metadata } from "next";
+"use client";
 import Link from "next/link";
-import { UserAuthForm } from "@/components/user-auth-form";
+import { ReactNode } from "react";
 
-export const metadata: Metadata = {
-  title: "Authentication",
-  description: "Authentication forms built using the components.",
-};
+interface AuthLayoutProps {
+  children: ReactNode;
+}
 
-export default function AuthenticationPage() {
+export default function AuthLayout({ children }: AuthLayoutProps) {
   return (
     <>
       <div className="container relative hidden flex-col items-center justify-center md:grid lg:max-w-none lg:grid-cols-2 lg:px-0">
@@ -40,11 +39,7 @@ export default function AuthenticationPage() {
         </div>
         <div className="lg:p-8">
           <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]">
-            <div className="flex flex-col space-y-2 text-center">
-              <h1 className="text-2xl font-semibold tracking-tight">Welcome Back</h1>
-              <p className="text-sm text-muted-foreground">Enter your email below to sign in</p>
-            </div>
-            <UserAuthForm />
+            {children}
             <p className="px-8 text-center text-sm text-muted-foreground">
               By clicking continue, you agree to our{" "}
               <Link href="/terms" className="underline underline-offset-4 hover:text-primary">
