+++ apps/buildog/src/app/(auth)/login/page.tsx
@@ -1,0 +1,19 @@
+import { Metadata } from "next";
+import { UserAuthForm } from "@/components/user-auth-form";
+
+export const metadata: Metadata = {
+  title: "Authentication",
+  description: "Authentication forms built using the components.",
+};
+
+export default function AuthenticationPage() {
+  return (
+    <>
+      <div className="flex flex-col space-y-2 text-center">
+        <h1 className="text-2xl font-semibold tracking-tight">Welcome Back</h1>
+        <p className="text-sm text-muted-foreground">Enter your email below to sign in</p>
+      </div>
+      <UserAuthForm />
+    </>
+  );
+}
